refactor: modularize CLI code for better maintainability
Break down the monolithic main.go (1400 lines) into modular packages:

Structure:
- internal/cli/commands/ - Command handlers (comments.go, notes.go)
- internal/cli/formatters/ - Output formatting (JSON, Toon, human-readable)
- internal/cli/helpers/ - Utility functions (SplitKeyValue, IntPtr)

Changes:
- main.go: Reduced from 1400 to 912 lines
- Moved comment commands to internal/cli/commands/comments.go (86 lines)
- Moved note commands to internal/cli/commands/notes.go (158 lines)
- Moved formatters to internal/cli/formatters/formatters.go (266 lines)
- Moved helpers to internal/cli/helpers/helpers.go (25 lines)
- Created unit tests for each module
- Removed main_test.go (tests moved to respective modules)

Benefits:
- Better code organization and separation of concerns
- Easier to test individual components
- Improved maintainability
- Clear module boundaries

All tests passing (31 tests across all packages)
